+++ controller/src/main/java/controller/controllers/StartController.java
@@ -1,34 +1,34 @@
-package controller.servlets;
+package controller.controllers;
 
-import classes.CommonSettings;
 import controller.dao.dto.TestsDTO;
 import controller.service.TestService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import controller.service.TestServiceImpl;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.context.support.SpringBeanAutowiringSupport;
+import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.List;
 
 
-@WebServlet("/start")
-public class StartServlet extends HttpServlet {
+@Controller
+public class StartController extends HttpServlet {
 
-    private  TestService testService;
+    private TestService testService;
 
-
-    @Override
-    public void init(ServletConfig config) throws ServletException {
-        super.init(config);
-        SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(this);
+    @ModelAttribute("userId")
+    public Integer userIdFromSession()
+    {
     }
 
     @Autowired
@@ -36,13 +36,8 @@ public class StartServlet extends HttpServlet {
         this.testService = testService;
     }
 
-
-
-    @Override
-
-   // @RequestMapping(value = "/start", method = RequestMethod.GET)
-
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    @RequestMapping(value = "/start", method = RequestMethod.GET)
+    public ModelAndView start(@RequestParam(name = "session") HttpSession session) {
         //FOR TEST ONLY!!!
         // int userId = (Integer) req.getSession().getAttribute(CommonSettings.AUTH_USER_ATTRIBUTE);
 
@@ -53,8 +48,4 @@ public class StartServlet extends HttpServlet {
         req.getRequestDispatcher("/controller.jsp").forward(req, resp);
     }
 
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
-    }
 }
