+++ controller/src/main/webapp/WEB-INF/web.xml
@@ -14,11 +14,7 @@
         <param-value>WEB-INF/mvc-dispatcher-servlet.xml</param-value>
     </context-param>
 
-
-
-
-
-<!--    <servlet>
+   <servlet>
         <servlet-name>mvc-dispatcher</servlet-name>
         <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
         <load-on-startup>1</load-on-startup>
@@ -26,7 +22,7 @@
     <servlet-mapping>
         <servlet-name>mvc-dispatcher</servlet-name>
         <url-pattern>/</url-pattern>
-    </servlet-mapping>-->
+    </servlet-mapping>
 
     
 </web-app>
